HTMLEditor: Add AI integration warnings & error; disable UI on error/warnings
