+++ packages/devextreme/js/__internal/ui/html_editor/utils/ai.ts
@@ -263,3 +263,5 @@ export const buildAICommandParams = <T extends AICommandNameExtended>(
     }
   }
 };
+
+export const containsCustomCommandWithoutPrompt = (commandsMap: CommandsMap): boolean => Object.keys(commandsMap).some((command) => command.startsWith('custom') && !commandsMap[command].prompt);
