+++ packages/devextreme/js/ui/widget/ui.errors.js
@@ -365,4 +365,12 @@ export default errorUtils(errors.ERROR_MESSAGES, {
      * @name ErrorsUIWidgets.W1025
      */
     W1025: '\'scrolling.mode\' is set to \'virtual\' or \'infinite\'. Specify the height of the component.',
+    /**
+     * @name ErrorsUIWidgets.W1026
+     */
+    W1026: 'AI feature is enabled. Pass the "aiIntegration" property.',
+    /**
+     * @name ErrorsUIWidgets.W1027
+     */
+    W1027: 'Prompt should be specified for custom command.',
 });
