+++ packages/devextreme/testing/tests/DevExpress.ui.widgets.htmlEditor/htmlEditorParts/toolbarModule.tests.js
@@ -8,10 +8,8 @@ import 'ui/drop_down_button';
 
 import Toolbar from '__internal/ui/html_editor/modules/m_toolbar';
 import FormDialog from '__internal/ui/html_editor/ui/m_formDialog';
-import AIDialog from '__internal/ui/html_editor/ui/aiDialog';
 import { noop } from 'core/utils/common';
 import keyboardMock from '../../../helpers/keyboardMock.js';
-import { getMenuItems, openAIDialog, openAIToolbarMenu } from '../../../helpers/aiToolbarMenu.js';
 import fx from 'common/core/animation/fx';
 import errors from 'ui/widget/ui.errors';
 import localization from 'localization';
@@ -45,8 +43,6 @@ const TEXTEDITOR_INPUT_CLASS = 'dx-texteditor-input';
 const DROPDOWNEDITOR_ICON_CLASS = 'dx-dropdowneditor-icon';
 const LIST_ITEM_CLASS = 'dx-list-item';
 const POPUP_TITLE_CLASS = 'dx-popup-title';
-const MENU_ITEM_CLASS = 'dx-menu-item';
-const MENU_CLASS = 'dx-menu';
 
 const BOLD_FORMAT_CLASS = 'dx-bold-format';
 const SIZE_FORMAT_CLASS = 'dx-size-format';
@@ -73,56 +69,6 @@ const TABLE_OPERATIONS = [
     'deleteRow',
     'deleteTable'
 ];
-
-const defaultAITranslateOptions = [
-    'Arabic',
-    'Chinese',
-    'English',
-    'French',
-    'German',
-    'Japanese',
-    'Spanish'
-];
-
-const defaultAIChangeToneOptions = [
-    'Professional',
-    'Casual',
-    'Straightforward',
-    'Confident',
-    'Friendly'
-];
-
-const defaultAIChangeStyleOptions = [
-    'Formal',
-    'Informal',
-    'Technical',
-    'Business',
-    'Creative',
-    'Journalistic',
-    'Academic',
-    'Persuasive',
-    'Narrative',
-    'Expository',
-    'Descriptive',
-    'Conversational'
-];
-
-const defaultAICommands = [
-    'summarize',
-    'proofread',
-    'expand',
-    'shorten',
-    'changeStyle',
-    'changeTone',
-    'translate',
-    'askAI'
-];
-
-const defaultAIOptions = {
-    translate: defaultAITranslateOptions,
-    changeTone: defaultAIChangeToneOptions,
-    changeStyle: defaultAIChangeStyleOptions,
-};
 
 const simpleModuleConfig = {
     beforeEach: function() {
@@ -246,49 +192,6 @@ const dialogModuleConfig = {
         };
 
         this.formDialog = new FormDialog(this.$element, { container: this.$element, position: null });
-    },
-    afterEach: function() {
-        fx.off = false;
-    }
-};
-
-const dialogAIModuleConfig = {
-    beforeEach: function() {
-        fx.off = true;
-
-        this.$element = $('#htmlEditor');
-
-        this.quillMock = {
-            on: noop,
-            keyboard: {
-                addBinding: noop
-            },
-            getModule: (moduleName) => this.quillMock[moduleName],
-            getSelection: () => ({ index: 0, length: 0 }),
-            getText: () => 'Test',
-        };
-
-        this.options = {
-            editorInstance: {
-                NAME: 'dxHtmlEditor',
-                addCleanCallback: noop,
-                addContentInitializedCallback: noop,
-                $element: () => {
-                    return this.$element;
-                },
-                _createComponent: ($element, widget, options) => {
-                    return new widget($element, options);
-                },
-                on: noop,
-                option: noop,
-                aiIntegration: {},
-                showAIDialog: (config) => {
-                    return this._aiDialog.show(config);
-                }
-            }
-        };
-
-        this._aiDialog = new AIDialog(this.$element, {});
     },
     afterEach: function() {
         fx.off = false;
@@ -1435,232 +1338,6 @@ testModule('Toolbar dialogs', dialogModuleConfig, () => {
         const isFieldsVisible = Boolean($fieldInputs.length);
 
         assert.notOk(isFieldsVisible, 'insertTable dialog is hidden');
-    });
-});
-
-testModule('Toolbar AI menu', dialogAIModuleConfig, () => {
-    QUnit.test('Should pass correct payload to dialog on item click', function(assert) {
-        this.options.items = [{ name: 'ai', commands: ['summarize'] }];
-        new Toolbar(this.quillMock, this.options);
-
-        const showSpy = sinon.spy(this.options.editorInstance, 'showAIDialog');
-
-        openAIDialog(this.$element);
-
-        assert.ok(showSpy.calledOnce, 'showAIDialog called');
-        assert.deepEqual(showSpy.firstCall.args[0], {
-            currentCommand: 'summarize',
-            currentCommandOption: undefined,
-            text: 'Test',
-            prompt: undefined,
-            commandsMap: {
-                summarize: {
-                    id: 'summarize',
-                    name: 'summarize',
-                    options: undefined,
-                    text: 'Summarize',
-                },
-            },
-        }, 'Correct config passed to dialog');
-    });
-
-    QUnit.test('Should pass correct payload to dialog on item click if there is custom command', function(assert) {
-        const prompt = () => 'custom prompt';
-        this.options.items = [{ name: 'ai', commands: [{ name: 'custom', prompt }] }];
-
-        new Toolbar(this.quillMock, this.options);
-
-        const showSpy = sinon.spy(this.options.editorInstance, 'showAIDialog');
-
-        openAIDialog(this.$element);
-
-        assert.ok(showSpy.calledOnce, 'showAIDialog called');
-        assert.deepEqual(showSpy.firstCall.args[0], {
-            currentCommand: 'custom0',
-            currentCommandOption: undefined,
-            text: 'Test',
-            prompt,
-            commandsMap: {
-                custom0: {
-                    id: 'custom0',
-                    name: 'custom',
-                    options: undefined,
-                    text: 'Custom',
-                    prompt,
-                },
-            },
-        }, 'Correct config passed to dialog');
-    });
-
-    QUnit.test('Should pass correct payload to dialog on item click if there is custom command with options', function(assert) {
-        const prompt = (param) => `custom prompt with ${param}`;
-
-        this.options.items = [
-            {
-                name: 'ai',
-                commands: [
-                    {
-                        name: 'custom',
-                        options: ['option 1'],
-                        prompt,
-                    },
-                ],
-            },
-        ];
-
-        new Toolbar(this.quillMock, this.options);
-
-        const showSpy = sinon.spy(this.options.editorInstance, 'showAIDialog');
-
-        openAIDialog(this.$element);
-
-        assert.ok(showSpy.calledOnce, 'showAIDialog called');
-        assert.deepEqual(showSpy.firstCall.args[0], {
-            currentCommand: 'custom0',
-            currentCommandOption: 'Option 1',
-            text: 'Test',
-            prompt,
-            commandsMap: {
-                custom0: {
-                    id: 'custom0',
-                    name: 'custom',
-                    options: ['Option 1'],
-                    text: 'Custom',
-                    prompt,
-                },
-            },
-        }, 'Correct config passed to dialog');
-    });
-
-    QUnit.test('showAIDialog is not called on root menu item click (with submenu)', function(assert) {
-        this.options.items = [{ name: 'ai', commands: ['translate'] }];
-        new Toolbar(this.quillMock, this.options);
-
-        const showSpy = sinon.spy(this.options.editorInstance, 'showAIDialog');
-
-        openAIToolbarMenu(this.$element);
-
-        const $rootItem = $(`.${MENU_ITEM_CLASS}`).eq(1);
-        $rootItem.trigger('dxclick');
-
-        assert.strictEqual(showSpy.callCount, 0, 'showAIDialog is not called on root item click');
-    });
-
-    ['ai', { name: 'ai' }].forEach(item => {
-        const itemType = typeof item === 'object' ? 'object' : 'string';
-
-        QUnit.test(`default commands, default options, AI item is passed as ${itemType}`, function(assert) {
-            this.options.items = [item];
-            new Toolbar(this.quillMock, this.options);
-
-            const menuItems = getMenuItems(this.$element);
-            const commandNames = menuItems.map(command => command.id);
-
-            assert.deepEqual(commandNames, defaultAICommands, 'commands match default list');
-
-            Object.entries(defaultAIOptions).forEach(([commandName, defaultOptions]) => {
-                const command = menuItems.find((command) => command.id === commandName);
-                const options = command.items.map((option) => option.id);
-                assert.deepEqual(options, defaultOptions, `Options for "${commandName}" match defaults`);
-            });
-        });
-    });
-
-    QUnit.test('specific commands, default options', function(assert) {
-        this.options.items = [{ name: 'ai', commands: ['translate', 'changeStyle', 'changeTone'] }];
-
-        new Toolbar(this.quillMock, this.options);
-
-        const menuItems = getMenuItems(this.$element);
-        const commandNames = menuItems.map(command => command.id);
-
-        assert.strictEqual(menuItems.length, 3, 'only specified commands rendered');
-        assert.deepEqual(commandNames, ['translate', 'changeStyle', 'changeTone'], 'commands match specified list');
-
-        Object.entries(defaultAIOptions).forEach(([commandName, defaultOptions]) => {
-            const command = menuItems.find((command) => command.id === commandName);
-            const actualOptions = command.items.map((option) => option.id);
-            assert.deepEqual(actualOptions, defaultOptions, `Options for "${commandName}" match defaults`);
-        });
-    });
-
-    QUnit.test('specific commands, specific options', function(assert) {
-        const commandOptions = ['English', 'Spanish'];
-        this.options.items = [{ name: 'ai', commands: [{ name: 'translate', options: commandOptions }] }];
-
-        new Toolbar(this.quillMock, this.options);
-
-        const menuItems = getMenuItems(this.$element);
-        const commandNames = menuItems.map(command => command.id);
-
-        const translateCommand = menuItems.find(command => command.id === 'translate');
-        const translateOptions = translateCommand.items.map(option => option.id);
-
-        assert.deepEqual(commandNames, ['translate'], 'commands match specified list');
-        assert.deepEqual(translateOptions, commandOptions, 'only specified options rendered');
-    });
-
-    QUnit.test('empty commands', function(assert) {
-        this.options.items = [{ name: 'ai', commands: [] }];
-
-        new Toolbar(this.quillMock, this.options);
-
-        const menuItems = getMenuItems(this.$element);
-        const commandNames = menuItems.map(command => command.id);
-
-        assert.deepEqual(commandNames, [], 'empty commands list in menu');
-    });
-
-    QUnit.test('empty command options', function(assert) {
-        this.options.items = [{ name: 'ai', commands: [{ name: 'translate', options: [] }] }];
-
-        new Toolbar(this.quillMock, this.options);
-
-        const menuItems = getMenuItems(this.$element);
-
-        const translateCommand = menuItems.find(command => command.id === 'translate');
-        const translateOptions = translateCommand.items.map(option => option.id);
-
-        assert.deepEqual(translateOptions, [], 'empty command options in menu');
-    });
-
-    QUnit.test('custom command text', function(assert) {
-        this.options.items = [{ name: 'ai', commands: [{ name: 'summarize', text: 'Summarize name' }] }];
-
-        new Toolbar(this.quillMock, this.options);
-
-        openAIToolbarMenu(this.$element);
-
-        const $menuItem = $(`.${MENU_ITEM_CLASS}`).last();
-        assert.strictEqual($menuItem.text(), 'Summarize name', 'custom command text rendered in menu');
-    });
-
-    QUnit.test('custom command', function(assert) {
-        this.options.items = [{
-            name: 'ai',
-            commands: [{
-                name: 'custom',
-                text: 'Custom command',
-            }]
-        }];
-
-        new Toolbar(this.quillMock, this.options);
-
-        openAIToolbarMenu(this.$element);
-
-        const $menuItem = $(`.${MENU_ITEM_CLASS}`).last();
-        assert.strictEqual($menuItem.text(), 'Custom command', 'custom command rendered in menu');
-    });
-
-    QUnit.test('root menu item is disabled if commands list is empty', function(assert) {
-        this.options.items = [{ name: 'ai', commands: [] }];
-        new Toolbar(this.quillMock, this.options);
-
-        openAIToolbarMenu(this.$element);
-
-        const menuInstance = $(`.${MENU_CLASS}`).dxMenu('instance');
-
-        assert.strictEqual(menuInstance.option('disabled'), true, 'menu is disabled');
     });
 });
